# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is API documentation on DataBx Sumbing1 Microservices",
    "version": "1.0.0",
    "title": "DataBx Sumbing1 Microservices",
    "contact": {
      "email": "akbar.muhammadakbarmaulana@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    {
      "name": "utils",
      "description": "Dead simple utils that can be useful to interact with microservices"
    },
    {
      "name": "generate",
      "description": "Utils that can be useful to, guess, generate"
    }
  ],
  "paths": {
    "/convert/json-to-bson": {
      "post": {
        "tags": [
          "utils"
        ],
        "summary": "Convert body json to bson encoded in byte",
        "description": "Only accept json and only return bson",
        "operationId": "ConvertJSONtoBSON",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "description": "ConvertJSONtoBSON success, with given request",
            "content": {
              "application/bson": {}
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules",
            "content": {
              "application/bson": {}
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "application/bson": {}
            }
          }
        }
      }
    },
    "/convert/bson-to-json": {
      "post": {
        "tags": [
          "utils"
        ],
        "summary": "Convert body bson to json",
        "description": "Vice versa and self-explanatory",
        "operationId": "ConvertBSONtoJSON",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/BSON"
        },
        "responses": {
          "200": {
            "description": "ConvertBSONtoJSON success, with given request",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/convert/image-to-bson-with-account-id": {
      "post": {
        "tags": [
          "utils"
        ],
        "summary": "Convert Image on body req to bson with additional account id from param",
        "description": "Don't ask me again",
        "operationId": "ConvertImagetoBSONWithAccountId",
        "security": [
          {"basicAuth":  []}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/xClientId"
          },
          {
            "$ref": "#/components/parameters/AccountId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }

        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Image"
        },
        "responses": {
          "200": {
            "description": "ConvertBSONtoJSON success, with given request",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Check request, body not comply with request rules",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "You are not permitted!"
          },
          "500": {
            "description": "General error, system cannot tell you what wrongs",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/generate/uuid": {
      "get": {
        "tags": [
          "generate"
        ],
        "summary": "Easy return of Google UUID generator",
        "description": "Don't ask me again",
        "operationId": "GenerateUUID",
        "security": [
          {"basicAuth":  []}
        ],
        "responses": {
          "200": {
            "description": "GenerateUUID success, with given no request",
            "content": {
              "text/plain": {}
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7778/sumbing1/api/v1"
    }
  ],
  "components": {
    "requestBodies": {
      "JSON": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "BSON": {
        "required": true,
        "content": {
          "application/bson": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
      "Image": {
        "required": true,
        "content": {
          "image/png": {
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
      "xClientId": {
        "name": "X-Client-ID",
        "in": "header",
        "description": "Custom Id that defines you as caller",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z0-9]{32}"
        }
      },
      "AccountId": {
        "name": "account_id",
        "in": "header",
        "description": "Account id that accompanies image data",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "RequestId": {
        "name": "request_id",
        "in": "header",
        "description": "Request id that accompanies, what else request",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 36,
          "pattern": "[A-Za-z0-9]{8}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{4}-[A-Za-z0-9]{12}"
        }
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}